netdev-afxdp: Add interrupt mode netdev class.
The patch adds a new netdev class 'afxdp-nonpmd' to enable afxdp
interrupt mode. This is similar to 'type=afxdp', except that the
is_pmd field is set to false. As a result, the packet processing
is handled by main thread, not pmd thread. This avoids burning
the CPU to always 100% when there is no traffic.

Tested-at: https://travis-ci.org/williamtu/ovs-travis/builds/655885506
Signed-off-by: William Tu <u9012063@gmail.com>
---
v4:
  - Previously crash and fix it with qid = qid % netdev_n_txq(netdev)
    Now remove it because Ilya's fix:
      "dpif-netdev: Force port reconfiguration to change dynamic_txqs."